# functions/src/controllers/swagger/bnb.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Binance Smart Chain API",
    "version": "1.0.0",
    "description": "API for handling BSC wallet operations, including wallet creation, balance retrieval, and transaction processing."
  },
  "paths": {
...
    "/wallet/wbgl/balance/get": {
      "post": {
        "summary": "Retrieve WBGL-BNB balance",
        "operationId": "wbglBNBBalanceHandler",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "address": {
                    "type": "string"
                  }
                },
                "required": [
                  "address"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "WBGL-BNB balance retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "balance": {
                      "type": "object",
                      "properties": {
                        "usd": {
                          "type": "number"
                        },
                        "balance": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tx/wbgl/history/get": {
      "post": {
        "summary": "Retrieve WBGL BEP-20 transactions",
        "operationId": "WBGLBEP20Explorer",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "address": {
                    "type": "string"
                  }
                },
                "required": [
                  "address"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "WBGL BEP-20 transactions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tx": {
                      "type": "array"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/tx/wbgl/create": {
      "post": {
        "summary": "Send WBGL on BNB Chain",
        "operationId": "sendWBGLBNBHandler",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number"
                  },
                  "to": {
                    "type": "string"
                  },
                  "privateKey": {
                    "type": "string"
                  }
                },
                "required": [
                  "amount",
                  "to",
                  "privateKey"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transaction successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "txHash": {
                      "type": "string"
                    },
                    "fee": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}